# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Portal API
  version: 0.1.0
  description: |
    Portal — self-hosted media platform API.

    This document is the **source of truth** for the HTTP contract.
    Server stubs (Go) and the TypeScript client are both generated from it.

    ## Errors

    Error responses use the **RFC 7807** `application/problem+json` shape
    (`#/components/schemas/Problem`), and that is the only error shape on the
    wire. The stable `type` URI is both the i18n key and the machine-stable
    slug: a domain code becomes `<module>/<code>` with underscores hyphenated
    (`email_taken` → `account/email-taken`), which is what the frontend's
    problem catalog keys on. Transport-level codes that add nothing to the
    status line (`unauthorized`, `forbidden`, `bad_request`, `internal`,
    `not_found`) collapse to `about:blank`.

    > **Migration status (v1):** the RFC 7807 retrofit per [D-7]/[D-29] is
    > **complete** — every handler, the auth/RBAC middleware, the rate limiter
    > and the `/continue` route now answer with `Problem`. The legacy
    > `{code, message}` body (`#/components/schemas/Error`) is emitted nowhere
    > and referenced by no operation; the schema is kept `deprecated` so a
    > generated client that still names the type keeps compiling, and should be
    > dropped at the next breaking release.

    Auth is **local password** ([ADR-06](../docs/adr/06-local-auth-model.md));
...
servers:
...
  # in any spec viewer hit a 404 on everything.
...
paths:
...
  /people:
    get:
      tags: [people]
      summary: List the caller's people (cursor, ordered by name)
      operationId: listPeople
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: cursor, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - name: circle
          in: query
          required: false
          description: Restrict to one section. Omit for every circle.
          schema: { $ref: "#/components/schemas/PersonCircle" }
      responses:
        "200":
          description: A cursor page of people
          content:
            application/json:
              schema:
                type: object
                required: [people]
                properties:
                  people: { type: array, items: { $ref: "#/components/schemas/Person" } }
                  next_cursor: { type: string, nullable: true }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [people]
      summary: Add a person
      operationId: createPerson
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PersonWrite" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Person" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /people/upcoming-birthdays:
    get:
      tags: [people]
      summary: Upcoming birthdays in the owner's timezone (soonest first)
      operationId: upcomingBirthdays
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: days, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 366, default: 14 } }
      responses:
        "200":
          description: Upcoming birthdays
          content:
            application/json:
              schema:
                type: object
                required: [upcoming]
                properties:
                  upcoming: { type: array, items: { $ref: "#/components/schemas/UpcomingBirthday" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /people/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [people]
      summary: Fetch one person (404 for others' rows)
      operationId: getPerson
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: The person, content: { application/json: { schema: { $ref: "#/components/schemas/Person" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [people]
      summary: Update a person (birthday null clears; resets current/future notices)
      operationId: updatePerson
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PersonWrite" } } }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Person" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      tags: [people]
      summary: Delete a person (idempotent)
      operationId: deletePerson
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Life-stream home (SPEC-06) ───────────────────────────────────────────
...
  /imports/{id}:
    parameters: [{ $ref: "#/components/parameters/AssetID" }]
    get:
      tags: [comic]
      summary: Poll an import job
      description: |
        `status` walks `pending → uploaded → running → done | failed`. `report`
        carries a per-entry result once the job has run.
      operationId: getImport
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The import job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /imports/{id}/zip:
    parameters: [{ $ref: "#/components/parameters/AssetID" }]
    put:
      tags: [comic]
      summary: Upload the import zip
      description: |
        Raw `application/zip` body, capped server-side. Completing the upload
        enqueues `comic:import_zip` on the default queue — deliberately NOT the
        heavy pool, because the job polls the asset statuses its own
        `media:process_image` tasks produce and must not occupy a slot they need.
      operationId: uploadImportZip
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/zip:
            schema: { type: string, format: binary }
      responses:
        "202":
          description: Accepted — the import job is queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/UnprocessableEntity" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }

  # ══ Tenancy + platform ════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: |
        Shared secret (COMIC_SYNC_SECRET) for service-to-service calls from the
        Python scraper. These endpoints carry no user session, which is why each
        one also takes owner_id — see the SPEC-02 P1.8 sync flow.
  parameters:
    AssetID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Birthday:
      type: object
      required: [month, day]
      properties:
        month: { type: integer, minimum: 1, maximum: 12 }
        day: { type: integer, minimum: 1, maximum: 31 }
        year: { type: integer, nullable: true, description: Optional — many people won't share a year. }
        calendar: { type: string, enum: [solar, lunar], default: solar }
    Person:
      type: object
      required: [id, display_name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        relationship: { type: string, nullable: true }
        birthday:
          allOf: [{ $ref: "#/components/schemas/Birthday" }]
          nullable: true
        contact: { type: object, additionalProperties: true }
        note_md: { type: string, nullable: true }
        avatar_asset_id: { type: string, format: uuid, nullable: true }
        circle: { $ref: "#/components/schemas/PersonCircle" }
        linked_user_id:
          type: string
          format: uuid
          nullable: true
          description: |
            The portal account this entry stands for, when it stands for one —
            set by adding someone from `/people/suggestions`. Null for everyone
            who has no account here, which is most of a personal registry.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PersonCircle:
      type: string
      enum: [close_friend, family, other]
      default: other
      description: |
        Which fixed section of the people rail this person appears in
        (migration `0035_people_circles`).

        This is beside `relationship`, not instead of it: `relationship` stays
        the free text you wrote ("mẹ", "bạn đại học"), while `circle` is the
        closed set the UI can group by without guessing at what that text means.
    PersonWrite:
      type: object
      description: "POST requires display_name; PATCH accepts any subset. A birthday of null clears it."
      properties:
        display_name: { type: string, minLength: 1, maxLength: 120 }
        relationship: { type: string, nullable: true }
        circle: { $ref: "#/components/schemas/PersonCircle" }
        linked_user_id:
          type: string
          format: uuid
          nullable: true
          description: On create only — records that this person is the given portal account.
        birthday:
          allOf: [{ $ref: "#/components/schemas/Birthday" }]
          nullable: true
        contact: { type: object, additionalProperties: true }
        note_md: { type: string, nullable: true }
    UpcomingBirthday:
      type: object
      required: [person_id, display_name, next_occurrence, days_until]
      properties:
        person_id: { type: string, format: uuid }
        display_name: { type: string }
        next_occurrence: { type: string, format: date }
        days_until: { type: integer }
        age_turning: { type: integer, nullable: true }

    # ── Life-stream home (SPEC-06) ──────────────────────────────────────────
...
    Problem:
      type: object
      description: |
        RFC 7807 problem detail, served as `application/problem+json`. `type` is a
        stable URI (also the i18n key); Portal adds the extensions below.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: "about:blank"
          description: Stable problem-type URI (i18n key).
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code, repeated for convenience.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: |
            Reserved. The machine-stable slug lives in `type`
            (`account/invalid-credentials`), so nothing emits `code` today —
            do not branch on it.
        required_acr:
          type: string
          description: For step-up (403) — the ACR the caller must satisfy [D-27].
        enrollment_url:
          type: string
          format: uri
          description: For MFA-enrollment (403) — where to enroll a second factor [D-28].
      additionalProperties: true

    # Legacy error body. Superseded by Problem and emitted by nothing — kept
    # only so generated clients naming the type still compile (see
    # info.description). Referenced by no operation; drop when clients allow.
...
    ImportJob:
      type: object
      required: [id, status, total, succeeded, failed, report, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        comic_id: { type: string, format: uuid, nullable: true, description: Set for a whole-comic import. }
        chapter_id: { type: string, format: uuid, nullable: true, description: Set for a single-chapter import. }
        status: { type: string, enum: [pending, uploaded, running, done, failed] }
        total: { type: integer, description: Entries the zip was found to contain. }
        succeeded: { type: integer }
        failed: { type: integer }
        upload_ref: { type: string, nullable: true, description: "Present while status is `pending` — PUT the zip to /imports/{id}/zip." }
        error: { type: string, nullable: true }
        report:
          type: array
          description: Per-entry outcome, populated once the job has run.
          items:
            type: object
            required: [name, ok]
            properties:
              name: { type: string }
              ok: { type: boolean }
              error: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ══ Tenancy ═══════════════════════════════════════════════════════════
...